Fix test failures caused by a patch for Transform visitors
In `TypeCheck::MakeImplicitCastsExplicit::postorder(IR::Expression *)`,
a `Cast` instruction is generated even if the inferred type
is equal to the actual type of the argument, but is represented
by another instance of the object because the types are compared
by their references. `inferTypeFromContext()` expects the parent
node of the currently visited node to be updated in order to return
the same type object for the right child after updating the left child.
This fix makes the result of the function stable so that it returns
the same type object for both children regardless of updating the
parent node.

This fixes the following tests:
 * p14_to_16/testdata/p4_14_samples/03-Gateway.p4
 * p14_to_16/testdata/p4_14_samples/axon.p4
 * p14_to_16/testdata/p4_14_samples/gateway2.p4
 * p14_to_16/testdata/p4_14_samples/gateway3.p4
 * p14_to_16/testdata/p4_14_samples/testgw.p4